pyproject.toml
[project]
name = "clipsapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "black>=25.11.0",
    "elevenlabs>=2.25.0",
    "fastapi[standard]>=0.123.5",
    "isort>=7.0.0",
    "moviepy>=2.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[tool.black]
line-length = 60
target-version = ['py37']
include = '\.pyi?$'

extend-exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 60
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 4
use_parentheses = true
ensure_newline_before_comments = true
